Refactor code to import parseLine and languagesAllowed from separate files
