+++ src/constants.js
@@ -1,0 +1,4 @@
+export const languagesAllowed = {
+  ja: 'JA',
+  zh: 'ZH',
+};
